# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Faed Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: size
          in: query
          schema:
            type: string
        - name: brand
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsResponse"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /supplier/products:
    get:
      operationId: listSupplierProducts
      tags: [supplier]
      summary: My products
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [supplier]
      summary: Upload product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /supplier/products/{id}:
    patch:
      operationId: updateProduct
      tags: [supplier]
      summary: Edit product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [supplier]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
...
    Product:
      type: object
      required: [id, name, price, originalPrice, quantity, category]
      properties:
        id:
          type: integer
        supplierId:
          type: integer
        supplierName:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        quantity:
          type: integer
        imageUrl:
          type: string
          nullable: true
        discountReason:
          type: string
          nullable: true
        sizes:
          type: array
          items:
            type: string
        brand:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
    ProductsResponse:
      type: object
      required: [products, total]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
...
    ProductInput:
      type: object
      required: [name, category, price, originalPrice, quantity]
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        originalPrice:
          type: number
        quantity:
          type: integer
        imageUrl:
          type: string
          nullable: true
        discountReason:
          type: string
          nullable: true
        sizes:
          type: array
          items:
            type: string
        brand:
          type: string
          nullable: true
...