# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Faed Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  category: string;
  /** @nullable */
  description?: string | null;
  price: number;
  originalPrice: number;
  quantity: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  discountReason?: string | null;
  sizes?: string[];
  /** @nullable */
  brand?: string | null;
}